Create Movement 3 (MOSTLY FUNCTIONAL BUT NOT REALLY)
# note: in this programs current state, child objects for the player object must be made named 'wallCheckLeft' and 'wallCheckRight' and moved to the left and right of the players hitbox.

# alternitively, triggers can be placed on walls which, assuming the code is functional, would also allow you to wall jump, assuming you gave the wall a 'Wall' tag (NOT IDEAL, USE CHILD OBJECTS INSTEAD)

# the 'IsTouchingWall' command should be tested to see if a Physics2D.OverlapBox could be used instead, as it would probably work better.

# in this program's current state, it is unsure whether either of the wall jump codes work, and it is definetly sure the boost code doesn't work, as I don't know how to check if the player is holding the left shift key

overall the functions a mess but I'm writing it on a computer w/o unity so I can't really test what works and what doesn't so I'm just coding whatever sounds like it will work. Compared to 'Movement 2 (NOT FUNCTIONAL)' this one is probably more not functional, but with the grass roots for a wall jump ability that doesn't interfere with the double jump, and doesn't work when touching the ground.