Create death detection 3
Respawn triggers work just as intended, only problem is that respawn locations are not based off the trigger touched, but just a counter that goes up upon entering a respawn trigger. benefit of this is that triggers don't have to be linked to respawn locations, streamlining level design (don't have to modify the function / put fields into the death function for different respawn amounts in different levels / rooms) but the downside is that respawns will always go in order, meaning that if the player skips a trigger then hits the next, they will still be sent back to the last triggers respawn point. To be fixed in a future script.